requirements.txt
Flask
Gunicorn
selenium
webdriver-manager
python-dotenv